+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Product price discount condition `hasBestPrice`.
+
 ## [2.6.0] - 2022-07-19
 ### Added
 - ConditionLayoutTelemarketing component.